+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend.go
@@ -6,8 +6,11 @@ import (
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_network_allocator"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider"
-	port_spec2 "github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/stacktrace"
 	"net"
 	"strconv"
@@ -47,23 +50,25 @@ var isContainerRunningDeterminer = map[types.ContainerStatus]bool{
 }
 
 // Unfortunately, Docker doesn't have an enum for the protocols it supports, so we have to create this translation map
-var portSpecProtosToDockerPortProtos = map[port_spec2.PortProtocol]string{
-	port_spec2.PortProtocol_TCP:  "tcp",
-	port_spec2.PortProtocol_SCTP: "sctp",
-	port_spec2.PortProtocol_UDP:  "udp",
+var portSpecProtosToDockerPortProtos = map[port_spec.PortProtocol]string{
+	port_spec.PortProtocol_TCP:  "tcp",
+	port_spec.PortProtocol_SCTP: "sctp",
+	port_spec.PortProtocol_UDP:  "udp",
 }
 
 type DockerKurtosisBackend struct {
 	dockerManager *docker_manager.DockerManager
 
+	dockerNetworkAllocator *docker_network_allocator.DockerNetworkAllocator
+
 	objAttrsProvider object_attributes_provider.DockerObjectAttributesProvider
 }
 
-func NewDockerKurtosisBackend(
-	dockerManager *docker_manager.DockerManager,
-) *DockerKurtosisBackend {
+func NewDockerKurtosisBackend(dockerManager *docker_manager.DockerManager) *DockerKurtosisBackend {
+	dockerNetworkAllocator := docker_network_allocator.NewDockerNetworkAllocator(dockerManager)
 	return &DockerKurtosisBackend{
 		dockerManager:    dockerManager,
+		dockerNetworkAllocator: dockerNetworkAllocator,
 		objAttrsProvider: object_attributes_provider.GetDockerObjectAttributesProvider(),
 	}
 }
@@ -109,7 +114,7 @@ func (backendCore *DockerKurtosisBackend) GetEnginePublicIPAndPort(
 // ====================================================================================================
 //                                     Private Helper Methods
 // ====================================================================================================
-func transformPortSpecToDockerPort(portSpec *port_spec2.PortSpec) (nat.Port, error) {
+func transformPortSpecToDockerPort(portSpec *port_spec.PortSpec) (nat.Port, error) {
 	portSpecProto := portSpec.GetProtocol()
 	dockerProto, found := portSpecProtosToDockerPortProtos[portSpecProto]
 	if !found {
@@ -167,9 +172,9 @@ func buildCombinedError(errorsById map[string]error, titleStr string) error {
 }
 
 
-func getPublicPortBindingFromPrivatePortSpec(privatePortSpec *port_spec2.PortSpec, allHostMachinePortBindings map[nat.Port]*nat.PortBinding) (
+func getPublicPortBindingFromPrivatePortSpec(privatePortSpec *port_spec.PortSpec, allHostMachinePortBindings map[nat.Port]*nat.PortBinding) (
 	resultPublicIpAddr net.IP,
-	resultPublicPortSpec *port_spec2.PortSpec,
+	resultPublicPortSpec *port_spec.PortSpec,
 	resultErr error,
 ) {
 	portNum := privatePortSpec.GetNumber()
@@ -231,10 +236,20 @@ func getPublicPortBindingFromPrivatePortSpec(privatePortSpec *port_spec2.PortSpe
 		)
 	}
 	hostMachinePortNumUint16 := uint16(hostMachinePortNumUint64) // Okay to do due to specifying the number of bits above
-	publicPortSpec, err := port_spec2.NewPortSpec(hostMachinePortNumUint16, portSpecProto)
+	publicPortSpec, err := port_spec.NewPortSpec(hostMachinePortNumUint16, portSpecProto)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred creating public port spec with host machine port num '%v' and protocol '%v'", hostMachinePortNumUint16, portSpecProto.String())
 	}
 
 	return hostMachineIp, publicPortSpec, nil
+}
+
+func getEnclaveIdFromNetwork(network *types.Network) (enclave.EnclaveID, error) {
+	labels := network.GetLabels()
+	enclaveIdLabelValue, found := labels[label_key_consts.EnclaveIDLabelKey.GetString()]
+	if !found {
+		return "", stacktrace.NewError("Expected to find network's label with key '%v' but none was found", label_key_consts.EnclaveIDLabelKey.GetString())
+	}
+	enclaveId := enclave.EnclaveID(enclaveIdLabelValue)
+	return enclaveId, nil
 }