+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -6,8 +6,11 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/stacktrace"
+	"io"
 )
 
 // TODO CALL THE METRICS LIBRARY EVENT-REGISTRATION FUNCTIONS HERE!!!!
@@ -60,15 +63,25 @@ func (backend *MetricsReportingKurtosisBackend) DestroyEngines(ctx context.Conte
 	return successes, failures, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) CreateEnclave(ctx context.Context, enclaveId string) (*enclave.Enclave, error) {
-	result, err := backend.underlying.CreateEnclave(ctx, enclaveId)
+func (backend *MetricsReportingKurtosisBackend) CreateEnclave(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	isPartitioningEnabled bool,
+)(*enclave.Enclave, error) {
+	result, err := backend.underlying.CreateEnclave(ctx, enclaveId, isPartitioningEnabled)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred creating enclave with ID '%v'", enclaveId)
+		return nil, stacktrace.Propagate(err, "An error occurred creating enclave with ID '%v' and is-partitioning-enabled value '%v'", enclaveId, isPartitioningEnabled)
 	}
 	return result, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) GetEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (map[string]*enclave.Enclave, error) {
+func (backend *MetricsReportingKurtosisBackend) GetEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+)(
+	map[enclave.EnclaveID]*enclave.Enclave,
+	error,
+){
 	results, err := backend.underlying.GetEnclaves(ctx, filters)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting enclaves using filters: %+v", filters)
@@ -76,7 +89,14 @@ func (backend *MetricsReportingKurtosisBackend) GetEnclaves(ctx context.Context,
 	return results, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) StopEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (successfulEnclaveIds map[string]bool, erroredEnclaveIds map[string]error, resultErr error) {
+func (backend *MetricsReportingKurtosisBackend) StopEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+)(
+	successfulEnclaveIds map[enclave.EnclaveID]bool,
+	erroredEnclaveIds map[enclave.EnclaveID]error,
+	resultErr error,
+){
 	successes, failures, err := backend.underlying.StopEnclaves(ctx, filters)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred stopping enclaves using filters: %+v", filters)
@@ -84,7 +104,14 @@ func (backend *MetricsReportingKurtosisBackend) StopEnclaves(ctx context.Context
 	return successes, failures, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) DestroyEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (successfulEnclaveIds map[string]bool, erroredEnclaveIds map[string]error, resultErr error) {
+func (backend *MetricsReportingKurtosisBackend) DestroyEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+)(
+	successfulEnclaveIds map[enclave.EnclaveID]bool,
+	erroredEnclaveIds map[enclave.EnclaveID]error,
+	resultErr error,
+) {
 	successes, failures, err := backend.underlying.DestroyEnclaves(ctx, filters)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying enclaves using filters: %+v", filters)
@@ -143,6 +170,245 @@ func (backend *MetricsReportingKurtosisBackend) DestroyAPIContainers(ctx context
 	successes, failures, err := backend.underlying.DestroyAPIContainers(ctx, filters)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying API containers using filters: %+v", filters)
+	}
+	return successes, failures, nil
+}
+
+
+func (backend *MetricsReportingKurtosisBackend) CreateModule(
+	ctx context.Context,
+	id module.ModuleID,
+	guid module.ModuleGUID,
+	containerImageName string,
+	serializedParams string,
+)(
+	newModule *module.Module,
+	resultErr error,
+) {
+	module, err := backend.underlying.CreateModule(
+		ctx,
+		id,
+		guid,
+		containerImageName,
+		serializedParams,
+		)
+	if err != nil {
+		return nil,
+		stacktrace.Propagate(
+			err,
+			"An error occurred creating module with ID '%v', GUID '%v', container image name '%v' and serialized params '%+v'",
+			id,
+			guid,
+			containerImageName,
+			serializedParams)
+	}
+
+	return module, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) GetModules(
+	ctx context.Context,
+	filters *module.ModuleFilters,
+)(
+	map[string]*module.Module,
+	error,
+) {
+	modules, err := backend.underlying.GetModules(ctx, filters)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting modules using filters: %+v", filters)
+	}
+	return modules, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) DestroyModules(
+	ctx context.Context,
+	filters *module.ModuleFilters,
+)(
+	successfulModuleIds map[module.ModuleGUID]bool,
+	erroredModuleIds map[module.ModuleGUID]error,
+	resultErr error,
+) {
+	successes, failures, err := backend.underlying.DestroyModules(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying modules using filters: %+v", filters)
+	}
+	return successes, failures, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) CreateUserService(
+	ctx context.Context,
+	id service.ServiceID,
+	guid service.ServiceGUID,
+	containerImageName string,
+	privatePorts []*port_spec.PortSpec,
+	entrypointArgs []string,
+	cmdArgs []string,
+	envVars map[string]string,
+	enclaveDataDirMntDirpath string,
+	filesArtifactMountDirpaths map[string]string,
+)(
+	newUserService *service.Service,
+	resultErr error,
+) {
+	userService, err := backend.underlying.CreateUserService(
+		ctx,
+		id,
+		guid,
+		containerImageName,
+		privatePorts,
+		entrypointArgs,
+		cmdArgs,
+		envVars,
+		enclaveDataDirMntDirpath,
+		filesArtifactMountDirpaths,
+		)
+	if err != nil {
+		return nil,
+		stacktrace.Propagate(
+			err,
+			"An error occurred creating the user service with ID '%v' and GUID '%v' using image '%v' with private ports '%+v' with entry point args '%+v', command args '%+v', environment vars '%+v', enclave data mount dirpath '%v' and file artifacts mount dirpath '%v'",
+			id,
+			guid,
+			containerImageName,
+			privatePorts,
+			entrypointArgs,
+			cmdArgs,
+			envVars,
+			enclaveDataDirMntDirpath,
+			filesArtifactMountDirpaths,
+			)
+	}
+	return userService, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) GetUserServices(
+	ctx context.Context,
+	filters *service.ServiceFilters,
+)(
+	map[service.ServiceGUID]*service.Service,
+	error,
+){
+	services, err := backend.underlying.GetUserServices(ctx, filters)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting user services using filters '%+v'", filters)
+	}
+	return services, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) GetUserServiceLogs(
+	ctx context.Context,
+	filters *service.ServiceFilters,
+)(
+	map[service.ServiceGUID]io.ReadCloser,
+	error,
+) {
+	userServiceLogs, err := backend.underlying.GetUserServiceLogs(ctx, filters)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting user service logs using filters '%+v'", filters)
+	}
+	return userServiceLogs, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) RunUserServiceExecCommand (
+	ctx context.Context,
+	serviceGUID service.ServiceGUID,
+	commandArgs []string,
+)(
+	resultExitCode int32,
+	resultOutput string,
+	resultErr error,
+) {
+	exitCode, output, err := backend.underlying.RunUserServiceExecCommand(ctx, serviceGUID, commandArgs)
+	if err != nil {
+		return 0, "", stacktrace.Propagate(
+			err,
+			"An error occurred running user service exec command with user service GUID '%v' and command args '%+v'",
+			serviceGUID,
+			commandArgs,
+			)
+	}
+	return exitCode, output, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) WaitForUserServiceHttpEndpointAvailability(
+	ctx context.Context,
+	serviceGUID service.ServiceGUID,
+	httpMethod string,
+	port uint32,
+	path string,
+	requestBody string,
+	bodyText string,
+	initialDelayMilliseconds uint32,
+	retries uint32,
+	retriesDelayMilliseconds uint32,
+)(
+	resultErr error,
+) {
+	if err := backend.underlying.WaitForUserServiceHttpEndpointAvailability(
+		ctx,
+		serviceGUID,
+		httpMethod,
+		port,
+		path,
+		requestBody,
+		bodyText,
+		initialDelayMilliseconds,
+		retries,
+		retriesDelayMilliseconds,
+		); err != nil {
+		return stacktrace.Propagate(
+			err,
+			"An error occurred waiting for http endpoint with path '%v', port '%v', request body '%v', body text '%v' from service with GUID '%v' to become available after '%v' retries and '%v' milliseconds between retries,",
+			path,
+			port,
+			requestBody,
+			bodyText,
+			serviceGUID,
+			retries,
+			retriesDelayMilliseconds,
+		)
+	}
+	return nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) GetShellOnUserService(
+	ctx context.Context,
+	serviceGUID service.ServiceGUID,
+)(
+	resultErr error,
+) {
+	if err := backend.underlying.GetShellOnUserService(ctx, serviceGUID); err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting shell on user service with GUID '%v'", serviceGUID)
+	}
+	return nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) StopUserServices(
+	ctx context.Context,
+	filters *service.ServiceFilters,
+)(
+	successfulUserServiceIds map[service.ServiceGUID]bool,
+	erroredUserServiceIds map[service.ServiceGUID]error,
+	resultErr error,
+) {
+	successes, failures, err := backend.underlying.StopUserServices(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred stopping user services using filters: %+v", filters)
+	}
+	return successes, failures, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) DestroyUserServices(
+	ctx context.Context,
+	filters *service.ServiceFilters,
+)(
+	successfulUserServiceIds map[service.ServiceGUID]bool,
+	erroredUserServiceIds map[service.ServiceGUID]error,
+	resultErr error,
+) {
+	successes, failures, err := backend.underlying.DestroyUserServices(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying user services using filters: %+v", filters)
 	}
 	return successes, failures, nil
 }