+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -5,7 +5,10 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"io"
 )
 
 // KurtosisBackend abstracts a Kurtosis backend, which will be a container engine (Docker or Kubernetes).
@@ -55,7 +58,8 @@ type KurtosisBackend interface {
 	// Creates an enclave with the given enclave ID
 	CreateEnclave(
 		ctx context.Context,
-		enclaveId string,
+		enclaveId enclave.EnclaveID,
+		isPartitioningEnabled bool,
 	) (
 		*enclave.Enclave,
 		error,
@@ -66,7 +70,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		map[string]*enclave.Enclave,
+		map[enclave.EnclaveID]*enclave.Enclave,
 		error,
 	)
 
@@ -77,8 +81,8 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		successfulEnclaveIds map[string]bool,
-		erroredEnclaveIds map[string]error,
+		successfulEnclaveIds map[enclave.EnclaveID]bool,
+		erroredEnclaveIds map[enclave.EnclaveID]error,
 		resultErr error,
 	)
 
@@ -87,10 +91,11 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		successfulEnclaveIds map[string]bool,
-		erroredEnclaveIds map[string]error,
+		successfulEnclaveIds map[enclave.EnclaveID]bool,
+		erroredEnclaveIds map[enclave.EnclaveID]error,
 		resultErr error,
 	)
+
 
 	CreateAPIContainer(
 		ctx context.Context,
@@ -108,7 +113,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *api_container.APIContainerFilters,
 	) (
-		// Matching API containers, keyed by their enclave ID
+	// Matching API containers, keyed by their enclave ID
 		map[string]*api_container.APIContainer,
 		error,
 	)
@@ -118,7 +123,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		// Successful & errored API containers are keyed by their enclave ID
+	// Successful & errored API containers are keyed by their enclave ID
 		successApiContainerIds map[string]bool,
 		erroredApiContainerIds map[string]error,
 		resultErr error,
@@ -129,10 +134,113 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		// Successful & errored API containers are keyed by their enclave ID
+	// Successful & errored API containers are keyed by their enclave ID
 		successApiContainerIds map[string]bool,
 		erroredApiContainerIds map[string]error,
 		resultErr error,
+	)
+
+	// Create a module from a container image with serialized params
+	CreateModule(
+		ctx context.Context,
+		id module.ModuleID,
+		guid module.ModuleGUID,
+		containerImageName string,
+		serializedParams string,
+	)(
+		newModule *module.Module,
+		resultErr error,
+	)
+
+	// Gets modules using the given filters, returning a map of matched modules identified by their module ID
+	GetModules(ctx context.Context, filters *module.ModuleFilters) (map[string]*module.Module, error)
+
+	// Destroys the modules with the given filters, regardless of if they're running or not
+	DestroyModules(
+		ctx context.Context,
+		filters *module.ModuleFilters,
+	) (
+		successfulModuleIds map[module.ModuleGUID]bool, // "set" of module IDs that were successfully destroyed
+		erroredModuleIds map[module.ModuleGUID]error, // "set" of module IDs that errored when destroying, with the error
+		resultErr error, // Represents an error with the function itself, rather than the modules
+	)
+
+	// Creates a user service inside an enclave with the given configuration
+	CreateUserService(
+		ctx context.Context,
+		id service.ServiceID,
+		guid service.ServiceGUID,
+		containerImageName string,
+		privatePorts []*port_spec.PortSpec,
+		entrypointArgs []string,
+		cmdArgs []string,
+		envVars map[string]string,
+		enclaveDataDirMntDirpath string,
+		filesArtifactMountDirpaths map[string]string,
+    )(
+		newUserService *service.Service,
+		resultErr error,
+	)
+
+	// Gets user services using the given filters, returning a map of matched user services identified by their ID
+	GetUserServices(ctx context.Context, filters *service.ServiceFilters) (map[service.ServiceGUID]*service.Service, error)
+
+	// Get user service logs using the given filters, returning a map of matched user services identified by their ID and a readCloser object for each one
+	GetUserServiceLogs(ctx context.Context, filters *service.ServiceFilters) (map[service.ServiceGUID]io.ReadCloser, error)
+
+	// Executes a shell command inside an user service instance indenfified by its ID
+	RunUserServiceExecCommand (
+		ctx context.Context,
+		serviceGUID service.ServiceGUID,
+		commandArgs []string,
+	)(
+		exitCode int32,
+		output string,
+		resultErr error,
+	)
+
+	// Wait for succesful http endpoint response which can be used to check if the service is available
+	WaitForUserServiceHttpEndpointAvailability(
+		ctx context.Context,
+		serviceGUID service.ServiceGUID,
+		httpMethod string, //The httpMethod used to execute the request. Valid values: GET and POST
+		port uint32, //The port of the service to check. For instance 8080
+		path string, //The path of the service to check. It mustn't start with the first slash. For instance `service/health`
+		requestBody string, //The content of the request body. Only valid when the httpMethod is POST
+		bodyText string, //If the endpoint returns this value, the service will be marked as available (e.g. Hello World).
+		initialDelayMilliseconds uint32, //The number of milliseconds to wait until executing the first HTTP call
+		retries uint32, //Max number of HTTP call attempts that this will execute until giving up and returning an error
+		retriesDelayMilliseconds uint32, //Number of milliseconds to wait between retries
+	)(
+		resultErr error,
+	)
+
+	// Get an interactive shell to execute commands in an user service
+	GetShellOnUserService(
+		ctx context.Context,
+		serviceGUID service.ServiceGUID,
+	)(
+		resultErr error,
+	)
+
+	// Stop user services using the given filters,
+	StopUserServices(
+		ctx context.Context,
+		filters *service.ServiceFilters,
+	)(
+		successfulUserServiceIds map[service.ServiceGUID]bool, // "set" of user service IDs that were successfully stopped
+		erroredUserServiceIds map[service.ServiceGUID]error, // "set" of user service IDs that errored when stopping, with the error
+		resultErr error, // Represents an error with the function itself, rather than the user services
+	)
+
+	// Destroy user services using the given filters,
+	DestroyUserServices(
+		ctx context.Context,
+		filters *service.ServiceFilters,
+	)(
+		successfulUserServiceIds map[service.ServiceGUID]bool, // "set" of user service IDs that were successfully destroyed
+		erroredUserServiceIds map[service.ServiceGUID]error, // "set" of user service IDs that errored when destroying, with the error
+		resultErr error, // Represents an error with the function itself, rather than the user services
 	)
 
 	// TODO CreateRepl
@@ -147,19 +255,4 @@ type KurtosisBackend interface {
 
 	// TODO RunReplExecCommand
 
-	// TODO CreateModule
-
-	// TODO DestroyModule
-
-	// TODO CreateUserService
-
-	// TODO GetUserServices
-
-	// TODO GetUserServiceLogs
-
-	// TODO StopUserServices
-
-	// TODO GetShellOnUserService
-
-	// TODO DestroyUserServices
 }