+++ container-engine-lib/lib/backend_interface/objects/enclave/enclave_filters.go
@@ -2,7 +2,7 @@ package enclave
 
 type EnclaveFilters struct {
 	// Disjunctive set of enclave IDs to operate on
-	IDs map[string]bool
+	IDs map[EnclaveID]bool
 
 	// Disjunctive set of enclave statuses to operate on
 	Statuses map[EnclaveStatus]bool