+++ container-engine-lib/lib/backend_interface/objects/module/module_filters.go
@@ -1,0 +1,7 @@
+package module
+
+type ModuleFilters struct {
+	// Disjunctive set of module GUIDs to find modules for
+	// If nil or empty, will match all GUIDs
+	GUIDs map[ModuleGUID]bool
+}